T3Controller: use spawn instead of fork
The previous implementation used fork to create subprocesses. This was quite
convenient, as all setup could be done in the master process, but due to the
way Python threads and fork() interact, could also cause the master process to
deadlock. Instead of relying on fork(), we explicitly send serialized versions
of both the global and job configs to a fresh subprocess on each job launch.