Add T3FilteringStockSelector
This implements stock selection based on the latest T2 results via a pipeline
that merges independent T2 docs by stock id. This is similar to, but more
efficient than, the transients->select->scienceRecords config from v0.6, since
it can be done e.g. before datapoints are loaded. As in v0.6, and Mongo query
expressions in general, you can only compare doc fields to a constant, not to
each other.

Correlated queries _can_ be implemented by prepending $addFields stages that
add the synthesized fields (e.g. difference of two chi^2 values), but a general
interface to support that is tricky.