T3Processor: explicitly mark runs as successful
Upon completion, T3Processor adds a sets the field `success` to `True` in the
run doc. By default, TimeLastRunModel will use this to ignore T3 processes that
are were interrupted by exceptions or killed before completing. Incomplete T3
runs can be considered by setting `require_success=False`.

This attempts to fix a long-standing annoyance with v0.6 T3s that communicated
with flaky upstream services (looking at you, MarshalPublisher). When these
services were unavailable, the T3s would fail, and would have to be configured
with a lookback time of e.g. 3 days to allow multiple attempts. While this
helped transients to be submitted to upstream services eventually, it also
always put the worst-case load on the database and T3 controller. With the
ability to ignore failed runs, processes can be configured with the nominal
lookback time, and will only extend this to cover previous runs if they did not
complete.