AmpelDB: get authentication from secret provider
The resource section of the config now stores only the URI of the mongo server, while authentication info is kept by the secret provider. This paves the way for dynamic account provisioning, where the symbolic role names in the db config map onto temporary users that may have access only to prefixed databases. Because AmpelDB defaults to an empty auth dict if no secret provider is configured, this works transparently for local testing databases with no authentication.