Allow tier-less processes
These are maintenance processes that are not directly involved in the Ampel data flow and so not bound to a particular tier, for example collecting metrics, posting exceptions to Slack, and eventually purging old stocks. Processes with tier=None are collected under process.ops in the config. AmpelController.get_processes() selects these if tier="ops". AmpelExceptionPublisher is now an instance of AbsOpsUnit, and invoked in the process ExceptionPublisher registered in the config.