T3ChannelProjector: add multi-channel projection
channel accepts the same schema as T3SimpleDataLoader.channel. While all the logical schemas are reduced to a set selection, this makes it possible to use the same schema to inform both T3SimpleDataLoader and T3ChannelProjector that you are only interested in documents related to a subset of channels.