AbsLegacyChannelTemplate: add t2 unit dependencies
Units configured in the `dependency` field of tied units are automatically added to the ingestion directive. This allows the configuration of said dependencies to be specified only once. Dependencies configured as defaults in unit definitions can't be automatically added because AbsLegacyChannelTemplate does not actually instantiate any units. A test case (marked xfail) exists to remind us of this case.