AbsLegacyChannelTemplate: do not auto request tied units ingestion
Partially reverts https://github.com/AmpelProject/Ampel-core/commit/3c4974e58a88737e44644512eea687f9b81e0402
The idea was defendable, to implement it right is nightmarish.
Let alone that is conflicts with "explicit is better than implicit".
Issues with the removed implementation:
- Results in erroneous config when hybrid tied t2s are used (or when 'link_override' is used)
- Resolving aliases fails (fixable easily for global aliases, with more work for scoped aliases)
There is no quick way to do this right and it may not be worth the effort.

This commit adds the following basic check:
if tied t2 units are requested in the model,
the units required by the defined tied units
must be present in the model as well.