AbsLegacyChannelTemplate: detect mismatched T2 configs
Since you have to type out the configuration twice for a tied unit, once
in the dependency, and once in the tied unit config, you're free to make
mistakes. AbsLegacyChannelTemplate now raises an error if a tied unit
config defines a t2_dependency that does not match any of the provided
unit configs.