DBContentLoader: ensure T2Documents are ordered by id
This ensures that the most recently created T2 document is at the end of
the list. When loading documents constrained to stock and channel, mongo
will use the stock/channel index. This means that the T2Documents
returned for a given stock are ordered by the [hash of the] channel id,
which for multi-channel selections means that the "most recent" result
will be only for the "larger" channel, even if a newer result is
available for the "smaller" channel.