T2Processor: preserve order of datapoints
Tied state units that depend on point units rely on datapoints being
presented in the same order as during ingestion, so e.g. `first` refers
to the same datapoint. If datapoints are loaded without sorting, they
are returned in order of _id, which for ZTF-IPAC alerts usually
increases monotonically for datapoints from a single stock, and so
matches the [explicitly sorted] compound payload.

When datapoints are duplicated, however, the ordering of datapoints is
ambiguous, and only preserved in the compound payload. This happens, for
example, when IPAC issues superseding alerts.