T0Compiler: handle distinct sets of dps in a cycle
When two channels are configured with the same shaper, but different
data sources (e.g. one direct ingestion directive, one multiplexed),
add() may be called multiple times with different `dps` for each value
of `channel`. Previously, `trace_id` was assumed to be a shorthand for
`dps`, and the actual value of `dps` was ignored if the `trace_id` was
already known in this cycle. Instead, keep a list of `(channels, dps)`
for each `trace_id`, and defer merging channel sets until `commit()`.