ChainedIngestionHandler: commit point t2s first
With this change, point (and stock) t2 shall be inserted before state t2s into the DB (for a given transient).
This order is better as state t2 can be bound to point t2s but the other way around.
'Shall' because we use bulk_writes with ordered=False for best performances.
Local tests in an unsharded env show that the order is respected though.