T3SequentialStager: remove cache capability
A better solution was implemented using T3PlainUnitExecutor in combination with T3AggregatingStager.
The main issue with the removed solution is that the input data is not checked (and is hardly checkable).
A cached result is loaded if the unit name and config (and possibly code) match.
This can yield erroneous behavior in situations where the input data changes
(a such change would not be detected and a potentially wrong cached result loaded).