T2Worker: reduce update latency
- the two updates to claim a t2 doc and write the result were done
with the default majority write concern.  while this was fine for
a standalone db backed by ssd storage, it introduced a full second
of latency to each op on the 3-member replica set i'm running for
elasticc.  now there's on option to only wait for the primary to
acknowledge the write in memory, which typically happens within 10
ms.  this decreases (wallclock) processing time by a factor ~2

- the stock document was updated after every t2, again waiting a
full second for the write to complete.  now there's an option to
bunch these into bulk updates like is done in AlertConsumer; like
AlertConsumer, it defaults to 500 updates.  this adds a factor ~2.5
speedup.

- gc.collect() was run after every single doc by default, and each
call took ~40x longer than running the actual t2 unit.  turning
this off yielded another speed up of ~8