Add basic stock purge (#192)
* AmpelDB: allow directConnection

* TimeConstraintModel: take now as a parameter

useful for calculating time constraints w.r.t. something other than the
system clock, e.g. the time when the latest stock was inserted

* Add MongoStockDeleter

Deletes stocks (and all associated t0/1/2 docs) based on a selection
similar to T3StockSelector. This is not quite a purge unit (the docs are
not archived anywhere), but can be used to keep to working set small.

* Bump version to 0.8.3b12