ThreadedViewGenerator: consume at most once
Since each call to __iter__ consumed the queue until the sentinel
was reached, a T3ReviewUnit that consumed the generator twice would
hang forever, as the stager would only fill the queue once.  This
was particularly hard to detect if the unit was tested with
SimpleViewGenerator, which could be iterated over an arbitrary
number of times.  To catch this case, use at-most-once semantics
for SimpleViewGenerator as well.