Robustify T2DocView config resolution in T3 (#205)
* T3StockSelector: handle mock client

* DevAmpelContext: load confids from the database

This allows e.g.  dynamically registered configs to be resolved in
T2DocView even if the AlertConsumer task is skipped.

* T3SimpleStager: resolve T2 configs

* ThreadedViewGenerator: deal with unit exceptions

When a unit raises an exception, the generator returned by
ThreadedViewGenerator.__init__ would exit instantly, causing the
wrapped JoinableQueue to...never join.  Wrap in a try/finally block
to ensure that the queue is always consumed.